# openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Occupy API",
		"description": "\nAPI for a social + bookings platform for small businesses.\n\n- **Auth**: JWT Bearer (`Authorization: Bearer <token>`)\n- **Pagination**: Cursor-based via `cursor` and `limit` query params; responses include `items[]` and `nextCursor`.\n- **Errors**: Standardized envelope `{ \"error\": { \"code\", \"message\", \"details\" } }`.\n\nSee **/redoc** for a compact reference.\n",
		"version": "0.1.0"
	},
	"paths": {
...
		"/api/v1/cliques/feed": {
			"get": {
				"tags": ["Cliques"],
				"summary": "View clique feed",
				"description": "Paginated feed of posts from cliques the user belongs to.",
				"operationId": "feed_api_v1_cliques_feed_get",
				"security": [{ "HTTPBearer": [] }, { "BearerAuth": [] }],
				"parameters": [
					{
						"name": "cursor",
						"in": "query",
						"required": false,
						"schema": {
							"anyOf": [{ "type": "string" }, { "type": "null" }],
							"title": "Cursor"
						}
					},
					{
						"name": "limit",
						"in": "query",
						"required": false,
						"schema": {
							"type": "integer",
							"maximum": 100,
							"minimum": 1,
							"default": 20,
							"title": "Limit"
						}
					},
					{ "$ref": "#/components/parameters/CursorParam" },
					{ "$ref": "#/components/parameters/LimitParam" }
				],
				"responses": {
					"200": {
						"description": "Feed page",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/CursorPagePosts" }
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			}
		},
		"/api/v1/cliques": {
			"post": {
				"tags": ["Cliques"],
				"summary": "Create clique",
				"description": "Business pages create cliques to organize services and bookings.",
				"operationId": "create_api_v1_cliques_post",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/CliqueCreate" }
						}
					},
					"required": true
				},
				"responses": {
					"201": {
						"description": "Clique created",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/Clique" }
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"422": {
						"description": "Unprocessable Content",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					}
				},
				"security": [{ "HTTPBearer": [] }, { "BearerAuth": [] }]
			}
		},
		"/api/v1/cliques/{cliqueId}": {
			"get": {
				"tags": ["Cliques"],
				"summary": "Get clique",
				"description": "Retrieve the public profile for a clique, with visibility rules applied.",
				"operationId": "get_clique_api_v1_cliques__cliqueId__get",
				"security": [{ "HTTPBearer": [] }],
				"parameters": [
					{
						"name": "cliqueId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid",
							"title": "Cliqueid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Clique profile",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": true,
									"title": "Response Get Clique Api V1 Cliques  Cliqueid  Get"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			},
			"patch": {
				"tags": ["Cliques"],
				"summary": "Update clique",
				"description": "Clique owners can edit name, imagery, privacy, and cancellation policy.",
				"operationId": "update_clique_api_v1_cliques__cliqueId__patch",
				"security": [{ "HTTPBearer": [] }, { "BearerAuth": [] }],
				"parameters": [
					{
						"name": "cliqueId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid",
							"title": "Cliqueid"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CliqueUpdate",
								"examples": {
									"adjust_policy": {
										"summary": "Update cancellation policy",
										"value": { "cancellationCutoffHours": 12 }
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Clique updated",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/Clique" }
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"422": {
						"description": "Unprocessable Content",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					}
				}
			},
			"delete": {
				"tags": ["Cliques"],
				"summary": "Delete clique",
				"description": "Clique owners can permanently delete their clique.",
				"operationId": "delete_clique_api_v1_cliques__cliqueId__delete",
				"security": [{ "HTTPBearer": [] }, { "BearerAuth": [] }],
				"parameters": [
					{
						"name": "cliqueId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid",
							"title": "Cliqueid"
						}
					}
				],
				"responses": {
					"204": { "description": "Clique deleted" },
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			}
		},
		"/api/v1/cliques/{cliqueId}/join": {
			"post": {
				"tags": ["Cliques"],
				"summary": "Request to join clique",
				"description": "Join a clique using an invite token when required.",
				"operationId": "join_api_v1_cliques__cliqueId__join_post",
				"security": [{ "HTTPBearer": [] }, { "BearerAuth": [] }],
				"parameters": [
					{
						"name": "cliqueId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid",
							"title": "Cliqueid"
						}
					},
					{
						"name": "inviteToken",
						"in": "query",
						"required": false,
						"schema": {
							"anyOf": [{ "type": "string" }, { "type": "null" }],
							"title": "Invitetoken"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Join request recorded",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/CliqueMember" }
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"422": {
						"description": "Unprocessable Content",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					}
				}
			}
		},
		"/api/v1/cliques/{cliqueId}/members/me": {
			"delete": {
				"tags": ["Cliques"],
				"summary": "Leave clique",
				"description": "Members can leave a clique they previously joined.",
				"operationId": "leave_api_v1_cliques__cliqueId__members_me_delete",
				"security": [{ "HTTPBearer": [] }, { "BearerAuth": [] }],
				"parameters": [
					{
						"name": "cliqueId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid",
							"title": "Cliqueid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Left clique",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": { "type": "string" },
									"title": "Response Leave Api V1 Cliques  Cliqueid  Members Me Delete"
								},
								"example": { "status": "left" }
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			}
		},
		"/api/v1/cliques/{cliqueId}/members": {
			"get": {
				"tags": ["Cliques"],
				"summary": "List clique members",
				"description": "Paginated list of members with role and status.",
				"operationId": "list_members_api_v1_cliques__cliqueId__members_get",
				"security": [{ "HTTPBearer": [] }, { "BearerAuth": [] }],
				"parameters": [
					{
						"name": "cliqueId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid",
							"title": "Cliqueid"
						}
					},
					{
						"name": "cursor",
						"in": "query",
						"required": false,
						"schema": {
							"anyOf": [{ "type": "string" }, { "type": "null" }],
							"title": "Cursor"
						}
					},
					{
						"name": "limit",
						"in": "query",
						"required": false,
						"schema": {
							"type": "integer",
							"maximum": 100,
							"minimum": 1,
							"default": 20,
							"title": "Limit"
						}
					},
					{ "$ref": "#/components/parameters/CursorParam" },
					{ "$ref": "#/components/parameters/LimitParam" }
				],
				"responses": {
					"200": {
						"description": "Members page",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CursorPageCliqueMembers"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			}
		},
		"/api/v1/cliques/{cliqueId}/members/pending": {
			"get": {
				"tags": ["Cliques"],
				"summary": "List pending membership requests",
				"description": "View join requests awaiting moderation by the clique owner.",
				"operationId": "list_pending_members_api_v1_cliques__cliqueId__members_pending_get",
				"security": [{ "HTTPBearer": [] }, { "BearerAuth": [] }],
				"parameters": [
					{
						"name": "cliqueId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid",
							"title": "Cliqueid"
						}
					},
					{
						"name": "cursor",
						"in": "query",
						"required": false,
						"schema": {
							"anyOf": [{ "type": "string" }, { "type": "null" }],
							"title": "Cursor"
						}
					},
					{
						"name": "limit",
						"in": "query",
						"required": false,
						"schema": {
							"type": "integer",
							"maximum": 100,
							"minimum": 1,
							"default": 20,
							"title": "Limit"
						}
					},
					{ "$ref": "#/components/parameters/CursorParam" },
					{ "$ref": "#/components/parameters/LimitParam" }
				],
				"responses": {
					"200": {
						"description": "Pending members",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CursorPageCliqueMembers"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			}
		},
		"/api/v1/cliques/{cliqueId}/members/{memberId}/approve": {
			"post": {
				"tags": ["Cliques"],
				"summary": "Approve membership request",
				"description": "Clique owners approve pending members.",
				"operationId": "approve_membership_api_v1_cliques__cliqueId__members__memberId__approve_post",
				"security": [{ "HTTPBearer": [] }, { "BearerAuth": [] }],
				"parameters": [
					{
						"name": "cliqueId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid",
							"title": "Cliqueid"
						}
					},
					{
						"name": "memberId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid",
							"title": "Memberid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Membership approved",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/CliqueMember" }
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"422": {
						"description": "Unprocessable Content",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					}
				}
			}
		},
		"/api/v1/cliques/{cliqueId}/members/{memberId}/reject": {
			"post": {
				"tags": ["Cliques"],
				"summary": "Reject membership request",
				"description": "Decline a pending membership request.",
				"operationId": "reject_membership_api_v1_cliques__cliqueId__members__memberId__reject_post",
				"security": [{ "HTTPBearer": [] }, { "BearerAuth": [] }],
				"parameters": [
					{
						"name": "cliqueId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid",
							"title": "Cliqueid"
						}
					},
					{
						"name": "memberId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid",
							"title": "Memberid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Membership rejected",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": { "type": "string" },
									"title": "Response Reject Membership Api V1 Cliques  Cliqueid  Members  Memberid  Reject Post"
								},
								"example": { "status": "rejected" }
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"422": {
						"description": "Unprocessable Content",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"CliqueMember": {
				"properties": {
					"id": { "type": "string", "format": "uuid", "title": "Id" },
					"cliqueId": {
						"type": "string",
						"format": "uuid",
						"title": "Cliqueid"
					},
					"userId": { "type": "string", "format": "uuid", "title": "Userid" },
					"role": { "$ref": "#/components/schemas/Role" },
					"status": { "$ref": "#/components/schemas/MembershipStatus" },
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"title": "Createdat"
					}
				},
				"type": "object",
				"required": ["id", "cliqueId", "userId", "role", "status", "createdAt"],
				"title": "CliqueMember"
			},
...
			"CursorPagePosts": {
				"properties": {
					"nextCursor": {
						"anyOf": [{ "type": "string" }, { "type": "null" }],
						"title": "Nextcursor"
					},
					"meta": {
						"anyOf": [
							{ "additionalProperties": true, "type": "object" },
							{ "type": "null" }
						],
						"title": "Meta"
					},
					"items": {
						"items": { "$ref": "#/components/schemas/Post" },
						"type": "array",
						"title": "Items"
					}
				},
				"type": "object",
				"required": ["items"],
				"title": "CursorPagePosts"
			},
...
			"HTTPValidationError": {
				"properties": {
					"detail": {
						"items": { "$ref": "#/components/schemas/ValidationError" },
						"type": "array",
						"title": "Detail"
					}
				},
				"type": "object",
				"title": "HTTPValidationError"
			},
...
			"ErrorEnvelope": {
				"type": "object",
				"required": ["error"],
				"properties": {
					"error": {
						"type": "object",
						"required": ["code", "message", "details"],
						"properties": {
							"code": { "type": "string", "example": "validation_error" },
							"message": {
								"type": "string",
								"example": "Invalid parent comment"
							},
							"details": {
								"type": "object",
								"additionalProperties": true,
								"example": {}
							}
						}
					}
				}
			}
		},
		"securitySchemes": {
			"HTTPBearer": { "type": "http", "scheme": "bearer" },
			"BearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		},
		"parameters": {
			"CursorParam": {
				"name": "cursor",
				"in": "query",
				"required": false,
				"schema": { "type": "string" },
				"description": "Opaque pagination cursor from previous response `nextCursor`."
			},
			"LimitParam": {
				"name": "limit",
				"in": "query",
				"required": false,
				"schema": {
					"type": "integer",
					"default": 20,
					"maximum": 100,
					"minimum": 1
				},
				"description": "Max items to return (default 20, max 100)."
			}
		},
...
	}
}